style(HSU): rename verification -> validation
