+++ packages/http-signature-utils/src/index.ts
@@ -2,9 +2,6 @@ export { createHeaders, Headers } from './utils/headers'
 export { generateJwk, JWK } from './utils/jwk'
 export { parseOrProvisionKey } from './utils/key'
 export { createSignatureHeaders } from './utils/signatures'
-export {
-  validateSignatureHeaders,
-  validateSignature
-} from './utils/verification'
+export { validateSignatureHeaders, validateSignature } from './utils/validation'
 export { generateTestKeys, TestKeys } from './test-utils/keys'
 export { RequestLike } from 'http-message-signatures'
