+++ packages/http-signature-utils/src/utils/validation.ts
